LayoutTests/media/media-source/media-source-append-before-last-range-no-quota-exceeded-expected.txt


EVENT(sourceopen)
EVENT(updateend)
EXPECTED (video.currentTime == '120') OK
Appending PTS=120
EVENT(updateend)
Appending PTS=121
EVENT(updateend)
Appending PTS=122
EVENT(updateend)
Appending PTS=123
EVENT(updateend)
Appending PTS=124
EVENT(updateend)
Appending PTS=125
EVENT(updateend)
Appending PTS=126
EVENT(updateend)
Appending PTS=127
EVENT(updateend)
Appending PTS=128
EVENT(updateend)
Appending PTS=129
EVENT(updateend)
Appending PTS=130
EVENT(updateend)
Appending PTS=131
EVENT(updateend)
Appending PTS=132
EVENT(updateend)
Appending PTS=133
EVENT(updateend)
Appending PTS=134
EVENT(updateend)
Appending PTS=135
EVENT(updateend)
Appending PTS=136
EVENT(updateend)
Appending PTS=137
EVENT(updateend)
Appending PTS=138
EVENT(updateend)
Appending PTS=139
EVENT(updateend)
Appending PTS=140
EVENT(updateend)
Appending PTS=141
EVENT(updateend)
Appending PTS=142
EVENT(updateend)
Appending PTS=143
EVENT(updateend)
Appending PTS=144
EVENT(updateend)
Appending PTS=145
EVENT(updateend)
Appending PTS=146
EVENT(updateend)
Appending PTS=147
EVENT(updateend)
Appending PTS=148
EVENT(updateend)
Appending PTS=149
EVENT(updateend)
Appending PTS=150
EVENT(updateend)
Appending PTS=151
EVENT(updateend)
Appending PTS=152
EVENT(updateend)
Appending PTS=153
EVENT(updateend)
Appending PTS=154
EVENT(updateend)
Appending PTS=155
EVENT(updateend)
Appending PTS=156
EVENT(updateend)
Appending PTS=157
EVENT(updateend)
Appending PTS=158
EVENT(updateend)
Appending PTS=159
EVENT(updateend)
Appending PTS=160
EVENT(updateend)
Appending PTS=161
EVENT(updateend)
Appending PTS=162
EVENT(updateend)
Appending PTS=163
EVENT(updateend)
Appending PTS=164
EVENT(updateend)
Appending PTS=165
EVENT(updateend)
Appending PTS=166
EVENT(updateend)
Appending PTS=167
EVENT(updateend)
Appending PTS=168
EVENT(updateend)
Appending PTS=169
EVENT(updateend)
Appending PTS=170
EVENT(updateend)
Appending PTS=171
EVENT(updateend)
Appending PTS=172
EVENT(updateend)
Appending PTS=173
EVENT(updateend)
Appending PTS=174
EVENT(updateend)
Appending PTS=175
EVENT(updateend)
Appending PTS=176
EXPECTED (exception == 'QuotaExceededError: The quota has been exceeded.') OK
EXPECTED (bufferedRanges() == '[ 120...176 ]') OK
EXPECTED (video.currentTime == '115') OK
Appending PTS=115
EVENT(updateend)
Appending PTS=116
EVENT(updateend)
Appending PTS=117
EVENT(updateend)
Appending PTS=118
EVENT(updateend)
EXPECTED (exception != 'QuotaExceededError: The quota has been exceeded.') OK
EXPECTED (bufferedRanges() == '[ 115...119, 120...146 ]') OK
END OF TEST
